Refactor: Implement Task creation logic using factory method
