--- src/main/java/Deadline.java
@@ -1,14 +1,0 @@
-public class Deadline extends Task{
-    private String dueDate;
-    public Deadline(String name, String dueDate) {
-        super(name);
-        this.dueDate = dueDate;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[D]%s (by %s)",
-                super.toString(),
-                this.dueDate);
-    }
-}
