+++ src/main/java/Gopher.java
@@ -1,15 +1,7 @@
 import java.util.Scanner;
-import java.lang.StringBuilder;
 import java.util.ArrayList;
 
 public class Gopher {
-    // Initialize the input reader
-    private final static Scanner inputReader = new Scanner(System.in);
-
-    // Tasks List Data to handle user input tasks
-    private final static ArrayList<Task> taskList = new ArrayList<>();
-    private static int currentTaskNumber = 0;
-
     // Common Interface elements for easy reuse
     private final static String gopherLogo = """
               ____             _
@@ -21,6 +13,12 @@ public class Gopher {
             """;
     private final static String horizontalSeparator = "==================================================";
 
+    // Initialize the input reader
+    private final static Scanner inputReader = new Scanner(System.in);
+
+    // Tasks ArrayList to store user input tasks
+    private final static ArrayList<Task> taskList = new ArrayList<>();
+
     // Show greeting message when user first enter the application
     private static void greet () {
         System.out.println(horizontalSeparator);
@@ -28,132 +26,28 @@ public class Gopher {
         System.out.println("Hello! I am Gopher.\nWhat can I do for you?\n");
     }
 
-    // Check if the user input command is invalid
-    private static boolean isUnknownCommand(String command) {
-        return (!command.equalsIgnoreCase("todo"))
-                && (!command.equalsIgnoreCase("deadline"))
-                && (!command.equalsIgnoreCase("event"));
-    }
-
     // Handle the logic when user input command to add a new task
-    private static void addTask (String input) throws UnknownCommandException,
-            EmptyTaskDescriptionException, MissingTokenException {
-        // Split user command into tokens
-        String[] tokens = input.split(" ");
-
-        // Determine if the user command is valid
-        String taskType = tokens[0];
-        if (isUnknownCommand(taskType)) {
-            throw new UnknownCommandException(taskType);
+    private static void addTask (String input)  {
+        try {
+            Task newTask = Task.of(input);
+            taskList.add(newTask);
+            System.out.println(horizontalSeparator);
+            System.out.println("Got it! I have added this task:\n" + newTask);
+            System.out.println(String.format("Now you have %d %s in the list",
+                    taskList.size(),
+                    taskList.size() == 1 ? "task" : "tasks"));
+            System.out.println(horizontalSeparator + "\n");
+        } catch (UnknownCommandException
+        | EmptyTaskDescriptionException
+        | MissingTokenException e) {
+            System.out.println(e.getMessage());
         }
-
-        // Determine if there's any missing task description
-        if (tokens.length < 2) {
-            throw new EmptyTaskDescriptionException(taskType);
-        }
-
-        StringBuilder taskName = new StringBuilder();
-
-        if (taskType.equalsIgnoreCase("todo")) {
-            for (int i = 1; i < tokens.length; i++) {
-                taskName.append(tokens[i]);
-                if (i < tokens.length - 1) {
-                    taskName.append(" ");
-                }
-            }
-            taskList.add(new ToDo(taskName.toString()));
-        } else if (taskType.equalsIgnoreCase("deadline")) {
-            StringBuilder dueDate = new StringBuilder();
-
-            int byTokenIndex = -1;
-            for (int i = 1; i < tokens.length; i++) {
-                if (tokens[i].equalsIgnoreCase("/by")) {
-                    byTokenIndex = i;
-                }
-            }
-
-            if (byTokenIndex == -1) {
-                throw new MissingTokenException(taskType, "/by");
-            }
-
-            for (int i = 1; i < byTokenIndex; i++) {
-                taskName.append(tokens[i]);
-                if (i < byTokenIndex - 1) {
-                    taskName.append(" ");
-                }
-            }
-
-            for (int i = byTokenIndex + 1; i < tokens.length; i++) {
-                dueDate.append(tokens[i]);
-                if (i < tokens.length - 1) {
-                    dueDate.append(" ");
-                }
-            }
-
-            taskList.add(new Deadline(taskName.toString(),
-                    dueDate.toString()));
-        } else if (taskType.equalsIgnoreCase("event")) {
-            StringBuilder startDate = new StringBuilder();
-            StringBuilder endDate = new StringBuilder();
-
-            int fromTokenIndex = -1;
-            int toTokenIndex = -1;
-            for (int i = 1; i < tokens.length; i++) {
-                if (tokens[i].equalsIgnoreCase("/from")) {
-                    fromTokenIndex = i;
-                }
-                if (tokens[i].equalsIgnoreCase("/to")) {
-                    toTokenIndex = i;
-                }
-            }
-
-            if (fromTokenIndex == -1) {
-                throw new MissingTokenException(taskType, "/from");
-            }
-
-            if (toTokenIndex == -1) {
-                throw new MissingTokenException(taskType, "/to");
-            }
-
-            for (int i = 1; i < fromTokenIndex; i++) {
-                taskName.append(tokens[i]);
-                if (i < fromTokenIndex - 1) {
-                    taskName.append(" ");
-                }
-            }
-
-            for (int i = fromTokenIndex + 1; i < toTokenIndex; i++) {
-                startDate.append(tokens[i]);
-                if (i < toTokenIndex - 1) {
-                    startDate.append(" ");
-                }
-            }
-
-            for (int i = toTokenIndex + 1; i < tokens.length; i++) {
-                endDate.append(tokens[i]);
-                if (i < tokens.length - 1) {
-                    endDate.append(" ");
-                }
-            }
-            taskList.add(new Event(taskName.toString(),
-                    startDate.toString(),
-                    endDate.toString()));
-        }
-
-        Task addedTask = taskList.get(currentTaskNumber);
-        currentTaskNumber++;
-        System.out.println(horizontalSeparator);
-        System.out.println("Got it! I have added this task:\n" + addedTask);
-        System.out.println(String.format("Now you have %d %s in the list",
-                currentTaskNumber,
-                currentTaskNumber == 1 ? "task" : "tasks"));
-        System.out.println(horizontalSeparator + "\n");
     }
 
     // List out all the existing tasks in the application
     private static void listTasks() {
         System.out.println(horizontalSeparator);
-        for (int i = 1; i <= currentTaskNumber; i++) {
+        for (int i = 1; i <= taskList.size(); i++) {
             int currentTaskIndex = i - 1;
             String message = String.format("%d. %s", i, taskList.get(currentTaskIndex));
             System.out.println(message);
@@ -186,13 +80,13 @@ public class Gopher {
     // Delete the task with the corresponding task number
     private static void deleteTask(int taskNumber) {
         int taskIndex = taskNumber - 1;
-        String deletedTask = taskList.get(taskIndex).toString();
-        taskList.remove(taskIndex);
-        currentTaskNumber--;
+
         System.out.println(horizontalSeparator);
         System.out.println("Noted. I've removed this task:");
-        System.out.println(deletedTask);
+        System.out.println(taskList.get(taskIndex));
         System.out.println(horizontalSeparator + "\n");
+
+        taskList.remove(taskIndex);
     }
 
     // Say goodbye to the user and exit the application
@@ -234,13 +128,7 @@ public class Gopher {
                 int taskNumber = Integer.parseInt(tokens[1]);
                 deleteTask(taskNumber);
             } else {
-                try {
-                    addTask(userInput);
-                } catch (UnknownCommandException
-                        | EmptyTaskDescriptionException
-                        | MissingTokenException e) {
-                    System.out.println(e.getMessage());
-                }
+                addTask(userInput);
             }
         }
